Rename 'bench' package to 'benchmark'
