+++ pkg/benchmark/main.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package bench
+package benchmark
 
 import (
 	"context"
